AutoGen/FinancialAnalysisAndCoding/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "financial_analysis"
version = "0.1.0"
description = "A financial analysis project using AutoGen."
authors = [{ name = "Yan", email = "your_email@example.com" }]
dependencies = [
    "autogen==0.2.14",
    "python-dotenv==1.0.1",
    "yfinance==0.2.36",
    "matplotlib==3.8.3",
    "pandas==2.2.1",
    "pydantic==2.6.3",
    "loguru==0.7.2"
]

[tool.setuptools]
packages = ["financial_analysis"]

[tool.pytest.ini_options]
testpaths = ["tests"]
